% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_direction.R
\name{change_direction}
\alias{change_direction}
\title{change direction function}
\usage{
change_direction(value)
}
\arguments{
\item{value}{a numeric vector of values.}
}
\value{
A modified vector with values 1,-1, or 0 depending on whether the direction of change
was increasing, decreasing or no change.
}
\description{
A function to indicate the direction of change in value between current value and the next one.
i.e. whether the current value is bigger, smaller or equal to the next value
}
