% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\name{lineplot}
\alias{lineplot}
\title{lineplot function}
\usage{
lineplot(Data, xaxis, vector1, vector2)
}
\arguments{
\item{Data}{data frame.}

\item{xaxis}{the column including values of xaxis, e.g. time variable.}

\item{vector1}{the column including the values of the first line of interest.}

\item{vector2}{the column including the values of the second line of interest.}
}
\value{
a line plot created using ggplot
}
\description{
a wrapper function for ggplot showing two lines in a line plot.
}
