% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accleration.R
\name{process_acceleration}
\alias{process_acceleration}
\title{process_acceleration function}
\usage{
process_acceleration(acceleration, window, Data)
}
\arguments{
\item{acceleration}{a numeric vector including acceleration values.}

\item{window}{an integer specifying the number of observations in each comparison window for the threshold_function.}

\item{Data}{a data frame where the output of the function will be stored in two new columns}
}
\value{
data frame with two new columns including the updated acceleration column and change direction
}
\description{
a function that iteratively apply threshold_function and change direction function on
a data frame
}
