% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresold_func.R
\name{threshold_func}
\alias{threshold_func}
\title{threshold function}
\usage{
threshold_func(value, threshold = 0.5)
}
\arguments{
\item{value}{a numeric vector of values.}

\item{threshold}{a numeric value indicating threshold.}
}
\value{
A modified vector with values updated based on the threshold condition
}
\description{
A function to modify values only if the difference between the current value and the next value exceeds a threshold
}
